Bug 797502 - (RTL) - Right edge of the reports are locked
Running gnucash in Hebrew, the right edge of the reports are locked
while scrolling down and do not scroll with the rest of the report.

Adding a left and right margin of 8px fixes this and I think looks
better as text is not immediately against window edge.

Keeping top and bottom margin at 0px prevents the scrollbar from
appearing in charts.