Bug798341 - Right align in Hebrew - Account screen
When Gnucash is run in Hebrew which is a right-to-left language, the
monetary columns on the accounts page are left aligned but should be
right aligned. This is down to GTK transposing the column alignment
when using an RTL language so detect this and align to the left before
GTK transposes it.