Bug 798266 - Dropdown picker menu for category does not appear
If the OS/Theme is not using overlay scrolling, the register combos's
will not pop. This is down to the popup calculated height not including
the space for the horizontal scrollbar.

To fix this test for the horizontal scrollbar having the style class
"overlay-indicator" and if so add the height of the horizontal
scrollbar to the popup calculated height.

This can be seen if in a terminal you...
export GTK_OVERLAY_SCROLLING=0
and then run gnucash.