[gnc-budget-view.c] gbv_get_accumulated_budget_amount returns unreversed amount
Previously it would return an appropriately signed amount for
display. However it is also used when accumulating amounts therefore
needs to be sign-reversed again.

This function will now return, for the particular account and period,
an unreversed accumulated budget amount. The sign reversal now takes
place only for display using xaccSPrintAmount in #L1112