Bug 798415 - Due Bills Reminder Dialog - Right align 'amount'
When Gnucash is run in Hebrew which is a right-to-left language, the
monetary columns on the Bills/Invoices due reminder dialog are left
aligned but should be right aligned. This is down to GTK transposing
the column alignment when using an RTL language so detect this and
align to the left before GTK transposes it.