Allow to select saved report on print invoice button.
Use case : I want to be able to select a saved report I made to print
my invoices without having to open the saved report, change the option
and choose the invoice number. I also want to be able to print some of
my invoices using a specific template I made for quotes.

This change opens up the saved invoice dialog when the print invoice
button is clicked (only if the settings are set for this) and when I
select the saved report, the invoice report is created with the given
defaults and the correct invoice number.