Bug 798375 - Transfer Account Selection not editable after typing error
This bug is a result of not using overlayed scrollbars. When the scroll
bars are not overlayed, additional space is required for the bar and on
some desktops the use of gtk_widget_set_size_request is needed to
specify the height of the pop up.

This is only required when the number of entries is low and so the pop
up height will be less than the available space.