Don't compile Register2 code if it is not being used.
The Registger and Register2 code were entangled in such a way that
changing the Register headers made the Register2 code not compile.
Fixing this was difficult and would undubtedly make the Register2
code even worse than it is already.  Most of it has been ifdeffed
out.  Some of the conflicts have been fixed but many are left.