Bug 798438 - on reconcile window, 'R' column sorting broken
Sorting on the 'R' column was enabled in version 3.0 by mistake and
never worked as expected. Due to the 'R' column using a function to
populate the model, the existing sorting method would not work as this
was achieved by changing the query.

To fix this add a sorting function to sort the model for the 'R' column.