Add functions to gnc-query-view to do scrolling to selection
Add the ability when a row is selected to scroll to that row so it is
always visible when reordering the list.

The first function, gnc_query_use_scroll_to_selection can enable or
disable this for the query view, the default is off.

Two functions do the scrolling, gnc_query_scroll_to_selection and
gnc_query_force_scroll_to_selection with the first one respecting the
value of use scrolling to selection whereas the second will always do
it.