Allow to select saved report on print invoice button
This change was inspired by mildred's #PR1247 and uses the new
GncReportCombo to allow selection of 'Saved Invoice reports' when the
invoice print button is used or when printing multiple invoices.

This change presents a dialog with the default for the report combo set
to the preference setting. This dialog has a 2 second timeout which
will stop if a key is pressed or combo popped so that a different
report template can be selected and used. There is also an 'OK' button
that stops the time out and prints and a 'Cancel' button which cancels
the print.