Bug 798448 - Column 'num' cannot be null
Convert libofx's data.check_number and data.reference_number to "" if
they are any case variation of "null".

Note that SQL backends will not permit the word NULL as a value for the
number field if it's entered from the GUI, but this can be worked around
by quoting it as in Scheme: 'Null *will* work.