[dialog-sx-since-last-run] Fix handling of last run dialog on file open
When the "since last run" notification dialog is disabled on file open, it
leaks a reference to GncSxInstanceModel and any creation errors won't be
displayed. Avoid returning from the function too early.

If the "since last run" dialog is required but cancelled then the list of
created transactions isn't freed. Free the list when the dialog is closed.