Bug 798573 - Tab Width Behaviour
Currently when the notebook tabs on the left or right, the space used
fluctuates depending on tab label width. This change sets the label
width to the preference setting when the tabs are on the left or right
but when top or bottom the width is set to the number of characters
when shorter than the preference setting so they take up less room.