Bug 794584 - Register not updated when scheduled transactions created
Force the registers to refresh if there are automaticly created
transactions. That will mean two refreshes if the SLR dialog needs to
display and the user clicks OK, but the alternative is somehow tracking
whether to refresh if the user cancels the dialog. This is simpler and
an extra refresh call is harmless.