Add ability to hide placeholder and hidden accounts to GAS
Change the GNCAccountSel widget to hide placeholder and hidden
accounts. By default these accounts are hidden, but with the use of a
secondary icon a popup menu is presented where the list can be
configured to show them. Also when the GNCAccountSel is set to a
specific account with with gnc_account_sel_set_account, the placeholder
and hidden properties are checked and appropriate changes made so the
account will be visible.