Changes to window menu entries
Changed to use two digits for predefined menu labels so they will be
from Window00Action to Window09Action and change action name to
WindowAction. Also change the maximum number of windows from a macro to
a 'constexpr'. The maximum number of windows is still 10 but could be
increased by changing this number and adding entries to the .ui file.

Also added an info dialogue to advise when limit reached, no new window
entries will be added to the Windows menu.