Change the way Aqbanking /Ofx menus are created.
Instead of having the aqbabking / ofx already defined build the main
menu entries when plugin added and the action entry when the signal
menu_changed is emitted.

This has the added benefit when the Gnucash is built without the
aqbanking / ofx plugins the menu entries do not need to be hidden.