Use GtkApplicationWindow as basis for GncMainWindow
Geert's changes...
As a GtkApplicationWindow also implements the GActionMap interface
it can be used directly instead of the GSimpleActionGroup that was
set up to maintain our actions.

As a GtkApplicationWindow is not a GSimpleActionGroup (the other
object to implement GActionMap) this also needed a minor change in
the function signature of gnc_plugin_set_actions_enabled.