Bug 403979 - Balance column shows only low order digits when too narrow
When the column/cell is too narrow to display all of the numbers it
will only display the low-order digits which can be confusing.

To over come this, when the column/cell is too narrow elipsize at
the end so you have a visual indication the cell is to narrow.