Bug 798653 - Schedule Calendar event description pop up window does not track mouse position
The reporter was on Ubuntu 22.04.1 LTS which has wayland as the default
display server. When tested on a Fedora 37 VM which was also using
wayland the following message was seen on the terminal window...

Gdk-Message: 15:23:51.862: Window 0x28b0de0 is a temporary window
without parent, application will not be able to position it on screen.

This fix sets the transient parent for the dense calendar pop up.