Bug 798681 - Previously imported investment income transactions may not be filtered.
Resequence process_investment_transaction so that the first split is
the primary asset account split (cash unless it's a reinvest) followed
by the security asset account if it's not income and finally the income
split for income or reinvest transactions.

Note that there's also a sign change on the income splits for
income and reinvest transactions: testing showed the signs to be
backwards.