Keep ui files for optional modules within the module
This goes for ofx and aqbanking. If either
module is not part of the build, the ui files
should not be part of the resource of gnucash.
It turns out each compilation unit can have
its proper resource though, so the build system
was changed to include the ui files in a local resource
for the aqbanking and ofx modules instead of to
the gnucash global resource.