Bug 798702 - Crash in gnc_plugin_page_focus_idle_destroy() closing...
a report before it completes.

The problem was that the progress bar runs the event loop and lets
tab get destroyed while the stream handler is running, taking the
page and html objects along with it. Then the stream handler returns
and the callers try to use the freed page and html objects. Crash.

To avoid this take out weak pointers on the page and html and only
use the page and html if the respective finalize functions haven't
NULLed the weak pointer.