[stock-txn-asst] Refactor gtest-assistant-stock-transaction.cpp
For better readability. Removes gratuitous parameterized test
suite because the StockAssistantModel is stateful and can't
use easyTestCases when it and the book are re-created for
each instance, as is the case when the easyTestCases is passed
as a parameter set to a test suite.

Also replaced the shared_ptr<QofBook> with a unique_ptr.