Only reset csv importer internal state if new file is selected
The changes from
https://github.com/Gnucash/gnucash/commit/8ed40078b04f345e7104945f6928067c227a7740
and
https://github.com/Gnucash/gnucash/commit/78c8e8f3afbc8b5e65024479d9b4e382f83f6f6a
caused the importers to reset everything in the preview page whenever opening this
page, including when navigating back from a future page. This can be very frustrating
if many settings were adjusted but one went back just to tweak a single one.

With these changes the importer will only be reset whenever the user selects a
different file in the file chooser. So even navigating back to the file chooser,
but not changing files will keep the import state.