Bug 798753 - UX improvements in new import dialog
Rename the dialog to not mention export since we don't use it for that.
Change the key-column names to be specific to the listview.
Add tooltips to the listviews to explain what they're for
Persist the selected imexporter and profile in the book's state file.
Stop the inappropriate "there are transactions do you want to import
them?" dialog (missing AWAIT_TRANSACTIONS in gnc_ab_import_context call).