Bug 798759 - Register widths for 'with sub-account' not saved
Register widths are saved in the .gcm file under sections based on the
account guid. When you have for example an 'Assets' register open and
the 'with sub-account' register page 'Assets+' open, they will both
save there settings to the same section as there is only one guid.

To fix this, add a '+' to the end of the guid for the sub-account
section as done for the title.