Modify commit de9c0eb to not use xaccAccountSetSortReversed
Account sort order / sort by and filter by are all stored in the
register with the account functions in the process of being depreciated
so instead get the sort order from a new value added to SplitRegister.

Also in the general ledger, there is no default account and so was
causing an error in the trace file.