[engine] Fix AddressSanitizer build failure
When building with `-fsanitize=address`, the compiler does not understand
anymore that *month and *year are always set in compute_monthyear and warns
that they might be used uninitialized.
For a normal build it can figure out that `g_return_if_fail()` never triggers.

Eliminate that early return and rename the function to make it clear that the
only caller is in a context where the term type will be GNC_TERM_TYPE_PROXIMO.