[cmake] Only add -Werror to CFLAGS on release builds
During development it is desirable not to have to make the code
warning free for every build, as that slows down the testing cycle.

So move the -Werror flag to the release configuration flags.
Debug and RelWithDebInfo builds will still show the warning,
but continue building.

Also fix the improper use of CMAKE_C_FLAGS in setting
CMAKE_C_FLAGS_RELEASE and CMAKE_CXX_FLAGS_RELEASE.