[csv-transactions-export] fix category in export with simple layout
If Trading accounts is enabled, and the transaction is a transfer
between two accounts of differing commodities, the CSV simple-layout
export would incorrectly show "--Split Transaction--" as the
category.

This commit changes the algorithm to ignore the trading splits when
determining the other split account, thereby returning a more useful
other account name.