[trep-engine.scm] upgrade engine to improve calculated-cells handling
calculated-cells was formerly a vector of RHS info - heading,
calculator-fn, subtotal? etc.

This upgrades it so that calculator-fn now accepts two arguments -
split and transaction-row? which is a bool. It uses a basic record
with version and list-of-cells, and each cell is a record (instead of
a vector) with relevant members.

This also enables the engine to handle previous calculated-cells --
note the income-gst-statement.scm will offer the previous one and will
be transparently upgraded to the above -- see the function
upgrade-to-calculated-cells-v2