XML backend: Write host name (if available) and PID into lock file
Write the host name (if `gethostname` is available) and the PID of the
GnuCash process into the XML backend lock file.

User story: I synchronize my GnuCash data across multiple computers
using a cloud file synchronization service and run GnuCash on all of
these computers. When I forget to close GnuCash on one computer and
try to launch it on another, the lock file prevents me from doing
that, but now I need to figure out which computer it's running on so I
can close it there. Writing the host name and PID into the lock file
make this easy to do, so I don't have to check each of my computers
one by one to figure out where it's running.