[gnc-budget-view.c] much simpler negation rule: always negate.
previously negation was dependent on the row type and account
type. this change will modify the rule to *always* negate the total
amount prior to rendering.

this is because budget amounts are now natural numbers, i.e. budget
income usually negative, budget expense usually positive. the total of
budget amounts in a period (asset+liability+income+expense+equity)
should be zero.

the compulsory negation will ensure the budget equation is now much
more intuitive, i.e. an income *adds* to the budget, and an
expense *deducts* from budget; the "remaining to budget" is thus
positive if there is unallocated budget.