Fix null pointer dereference in gnc_split_register_save_to_copy_buffer()
If the "transfer" cell has been changed while cutting or copying a split
then it'll dereference the NULL "ft" value.

Check that "ft" is not NULL. This is equivalent to the original behaviour
before cb273b13a1269cf089c27e21192da444297fe3ae when the number of splits
would have been determined to be 0 for the SCM_UNDEFINED value of
"trans_scm".