Initialise all column widths in set_dimensions_pass_two()
It shouldn't be possible for the number of columns to be different between
the header cursor and the other cursors but the return value from
g_table_index() is checked just in case this happens.

If this did happen, the width would be set using an uninitialised value
so initialise them all to 0.