Bug 799048 - Hover on tab not correct
When the account name is changed, an event is used to update the tool
tip on the register tab. If the account is moved, the same event is
triggered but as the name is the same, the tool tip is not updated.

Add a new function 'main_window_update_page_long_name' to only update
the long name and the tool tip and call it from the register event
after setting the name. Also remove the setting of the long name from
'main_window_update_page_name'.