Bug 798995 - Keystrokes ignored during ledger entry
Some themes add additional pixels to list rows and dialogs, in this
case with 'TraditionalOK' 2 pixels are added to each entry in the
list and 2 to the popup. This make the calculation for the pop up
height wrong and so when compared to the allocation height it is
different so tries to resolve by re-poping the list and this
cuases an endless loop.

To fix this, use the calculated height which works for Adwaita but
if this is different to the popup allocation height use that.