Bug 728875 - Back button does not work in QIF import assistant
The current logic has each 'prepare' function jump us around in the assistant
flow in order to skip pages that aren't needed.  But that logic only allows
forward movement, effectively breaking the back button.

This patch introduces a custom 'next page' function, which GtkAssistant
supports, along with a shift in the paradigm.  Instead of the prepare
function jumping around, there is now a deliberate 'should I skip?' function
that is used to avoid displaying a page.

This does not make the flow perfect; unexpected exceptions still end the
process (see bug 698804) and going back with loaded files feels a bit clunky.