Bug 799090 - Right Clicking scheduled transaction
If you have a highlighted SX and you right click on another SX the new,
edit, delete popup uses the highlighted SX not the one you clicked on.
This is misleading because the box appears at the right clicked
location, and it looks like that is the SX it will affect.

To fix this, check to see if SX right clicked on is selected, if not
unselect all SX's and select the one clicked on before popup.